api/requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
openai==0.26.5
python-dotenv==0.21.1